requirements.txt
keyboard==0.13.5
mouse==0.7.1
pynput==1.7.6
Pillow==10.1.0
pystray==0.19.4
